+++ src/main.ts
@@ -12,27 +12,13 @@ const main = () => {
 
   const tableContainer = document.getElementById('table-container');
 
-  const taskInput = document.getElementById('input-task');
-  const prioritySelect = document.getElementById('select-priority');
-  const deadlineInput = document.getElementById('input-deadline');
-
-  const taskError = document.getElementById('error-task');
-  const priorityError = document.getElementById('error-priority');
-  const deadlineError = document.getElementById('error-deadline');
-
-  const submit = document.getElementById('submit');
+  const formContainer = document.getElementById('form-container');
 
   if (
     !(filterSlectContainer instanceof HTMLElement) ||
     !(removeAllDialogContainer instanceof HTMLElement) ||
     !(tableContainer instanceof HTMLElement) ||
-    !(taskInput instanceof HTMLInputElement) ||
-    !(prioritySelect instanceof HTMLSelectElement) ||
-    !(deadlineInput instanceof HTMLInputElement) ||
-    !(taskError instanceof HTMLSpanElement) ||
-    !(priorityError instanceof HTMLSpanElement) ||
-    !(deadlineError instanceof HTMLSpanElement) ||
-    !(submit instanceof HTMLButtonElement)
+    !(formContainer instanceof HTMLElement)
   ) {
     console.error('要素が存在しないか、誤ったタグが指定されています。');
     return;
@@ -44,7 +30,7 @@ const main = () => {
 
   initTodoTable(tableContainer);
 
-  initTodoForm({ taskInput, prioritySelect, deadlineInput, taskError, priorityError, deadlineError, submit });
+  initTodoForm(formContainer);
 
   initTodoDB(db);
 };